Implement get_page_objects method
It seems that this method is actually implemented in prawn-templates.
Direct support and integration of templates has been removed some time
ago and moved to the prawn-templates gem.

prawn-templates has a get_page_objects method that does a bit more. For
prawn itself we can just map the kids array to identifieres.