Fix serialisation of literal strings and Time
According to the PDF spec, a literal string may contain any character
except unbalanced parentheses ( and ) and the backslash. To make it
easy, these three characters need to be escaped.

Additionally, the carriage return needs to be escaped as \\r, i.e.
a backslash followed by the letter r. If it is not escaped, a single
\r would be read as \n according to the spec. And a backslash followed
by carriage return is ignored when read. So the only way to do it is
to use \\r.