Improve PDF dicts serialization stability
This should help with document serialization stability. Order of entries
in a dict is irrelevant so it should not have any bearing on document
identity.

PS: PDF spec uses logical order of entries in examples but Adobe Acrobat
actually sorts entries alphabetically. For instance, the most logically
important entry `/Type` is often the last.